+++ packages/@markuplint/ml-spec/gen/global-attribute.data.ts
@@ -10,6 +10,7 @@ export const globalAttrs: Record<
 		attrs: [
 			'accesskey',
 			'autocapitalize',
+			'autocorrect',
 			'autofocus',
 			'contenteditable',
 			'dir',