+++ packages/@markuplint/pug-parser/src/parser.ts
@@ -118,6 +118,21 @@ class PugParser extends Parser<ASTNode> {
 					},
 				);
 			}
+			case 'BlockComment': {
+				const lastBlock = originNode.block.nodes.at(-1);
+				const endOffset = lastBlock ? lastBlock.endOffset : originNode.endOffset;
+				const token = this.sliceFragment(originNode.offset, endOffset);
+				return this.visitComment(
+					{
+						...token,
+						depth,
+						parentNode,
+					},
+					{
+						isBogus: false,
+					},
+				);
+			}
 			case 'Tag': {
 				const namespace = getNamespace(originNode.name, parentNamespace);
 
@@ -186,9 +201,18 @@ class PugParser extends Parser<ASTNode> {
 				);
 			}
 			default: {
+				let tokenIncludesFile = token;
+
+				if ('file' in originNode) {
+					const interval = originNode.file.column - originNode.endColumn;
+					const fileOffset = originNode.endOffset + interval;
+					const fileEndOffset = fileOffset + originNode.file.path.length;
+					tokenIncludesFile = this.sliceFragment(originNode.offset, fileEndOffset);
+				}
+
 				return this.visitPsBlock(
 					{
-						...token,
+						...tokenIncludesFile,
 						depth,
 						parentNode,
 						nodeName: originNode.type,