+++ packages/@markuplint/create-rule/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.9](https://github.com/markuplint/markuplint/compare/@markuplint/create-rule@4.7.8...@markuplint/create-rule@4.7.9) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/create-rule
+
+## [4.7.8](https://github.com/markuplint/markuplint/compare/@markuplint/create-rule@4.7.7...@markuplint/create-rule@4.7.8) (2024-10-14)
+
+**Note:** Version bump only for package @markuplint/create-rule
+
+## [4.7.7](https://github.com/markuplint/markuplint/compare/@markuplint/create-rule@4.7.6...@markuplint/create-rule@4.7.7) (2024-09-23)
+
+**Note:** Version bump only for package @markuplint/create-rule
+
+## [4.7.6](https://github.com/markuplint/markuplint/compare/@markuplint/create-rule@4.7.5...@markuplint/create-rule@4.7.6) (2024-09-02)
+
+**Note:** Version bump only for package @markuplint/create-rule
+
 ## [4.7.5](https://github.com/markuplint/markuplint/compare/@markuplint/create-rule@4.7.4...@markuplint/create-rule@4.7.5) (2024-06-25)
 
 **Note:** Version bump only for package @markuplint/create-rule