+++ packages/@markuplint/erb-parser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.6.9](https://github.com/markuplint/markuplint/compare/@markuplint/erb-parser@4.6.8...@markuplint/erb-parser@4.6.9) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/erb-parser
+
+## [4.6.8](https://github.com/markuplint/markuplint/compare/@markuplint/erb-parser@4.6.7...@markuplint/erb-parser@4.6.8) (2024-10-14)
+
+**Note:** Version bump only for package @markuplint/erb-parser
+
+## [4.6.7](https://github.com/markuplint/markuplint/compare/@markuplint/erb-parser@4.6.6...@markuplint/erb-parser@4.6.7) (2024-09-23)
+
+**Note:** Version bump only for package @markuplint/erb-parser
+
+## [4.6.6](https://github.com/markuplint/markuplint/compare/@markuplint/erb-parser@4.6.5...@markuplint/erb-parser@4.6.6) (2024-09-02)
+
+**Note:** Version bump only for package @markuplint/erb-parser
+
 ## [4.6.5](https://github.com/markuplint/markuplint/compare/@markuplint/erb-parser@4.6.4...@markuplint/erb-parser@4.6.5) (2024-06-25)
 
 **Note:** Version bump only for package @markuplint/erb-parser