+++ packages/@markuplint/esm-adapter/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.5](https://github.com/markuplint/markuplint/compare/@markuplint/esm-adapter@4.5.4...@markuplint/esm-adapter@4.5.5) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/esm-adapter
+
+## [4.5.4](https://github.com/markuplint/markuplint/compare/@markuplint/esm-adapter@4.5.3...@markuplint/esm-adapter@4.5.4) (2024-10-14)
+
+**Note:** Version bump only for package @markuplint/esm-adapter
+
+## [4.5.3](https://github.com/markuplint/markuplint/compare/@markuplint/esm-adapter@4.5.2...@markuplint/esm-adapter@4.5.3) (2024-09-23)
+
+**Note:** Version bump only for package @markuplint/esm-adapter
+
+## [4.5.2](https://github.com/markuplint/markuplint/compare/@markuplint/esm-adapter@4.5.1...@markuplint/esm-adapter@4.5.2) (2024-09-02)
+
+**Note:** Version bump only for package @markuplint/esm-adapter
+
 ## [4.5.1](https://github.com/markuplint/markuplint/compare/@markuplint/esm-adapter@4.5.0...@markuplint/esm-adapter@4.5.1) (2024-06-25)
 
 **Note:** Version bump only for package @markuplint/esm-adapter