+++ packages/@markuplint/html-spec/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.9.3](https://github.com/markuplint/markuplint/compare/@markuplint/html-spec@4.9.2...@markuplint/html-spec@4.9.3) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/html-spec
+
+## [4.9.2](https://github.com/markuplint/markuplint/compare/@markuplint/html-spec@4.9.1...@markuplint/html-spec@4.9.2) (2024-10-14)
+
+**Note:** Version bump only for package @markuplint/html-spec
+
+## [4.9.1](https://github.com/markuplint/markuplint/compare/@markuplint/html-spec@4.9.0...@markuplint/html-spec@4.9.1) (2024-09-23)
+
+**Note:** Version bump only for package @markuplint/html-spec
+
+# [4.9.0](https://github.com/markuplint/markuplint/compare/@markuplint/html-spec@4.8.2...@markuplint/html-spec@4.9.0) (2024-09-02)
+
+### Features
+
+- **html-spec:** add the `attributionsrc` attr as experimental to `a`, `img`, and `script` elements ([2f44802](https://github.com/markuplint/markuplint/commit/2f44802264e3b51fa8f264536637bf419c86ca05))
+- **html-spec:** remove `onredo` and `onundo` attributes from the `body` element ([b428fb5](https://github.com/markuplint/markuplint/commit/b428fb5fcd3e0e28d9ed85eb93cf6aad7f081942))
+- **html-spec:** remove attributes from the obsolete `menuitem` element ([32d8c41](https://github.com/markuplint/markuplint/commit/32d8c415b03fbe68d5513d16fcaebf7318d289bf))
+- **html-spec:** remove the `cols` attribute from the `pre` element ([7605d7d](https://github.com/markuplint/markuplint/commit/7605d7db3c9dd40e190b37580d6fb53c75cff692))
+- **html-spec:** remove the `manifest` attribute from the `html` element ([5954a44](https://github.com/markuplint/markuplint/commit/5954a4490018d178ae71324badfcf1a352d9b07f))
+- **html-spec:** remove the `methods` attribute from the `link` element ([481a9b4](https://github.com/markuplint/markuplint/commit/481a9b49bef3f5b5546f569c24dd956b5a3dcd54))
+
 ## [4.8.2](https://github.com/markuplint/markuplint/compare/@markuplint/html-spec@4.8.1...@markuplint/html-spec@4.8.2) (2024-06-25)
 
 ### Bug Fixes