+++ packages/@markuplint/htmx-parser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.7](https://github.com/markuplint/markuplint/compare/@markuplint/htmx-parser@4.0.6...@markuplint/htmx-parser@4.0.7) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/htmx-parser
+
+## [4.0.6](https://github.com/markuplint/markuplint/compare/@markuplint/htmx-parser@4.0.5...@markuplint/htmx-parser@4.0.6) (2024-10-14)
+
+**Note:** Version bump only for package @markuplint/htmx-parser
+
+## [4.0.5](https://github.com/markuplint/markuplint/compare/@markuplint/htmx-parser@4.0.4...@markuplint/htmx-parser@4.0.5) (2024-09-23)
+
+**Note:** Version bump only for package @markuplint/htmx-parser
+
+## [4.0.4](https://github.com/markuplint/markuplint/compare/@markuplint/htmx-parser@4.0.3...@markuplint/htmx-parser@4.0.4) (2024-09-02)
+
+**Note:** Version bump only for package @markuplint/htmx-parser
+
 ## [4.0.3](https://github.com/markuplint/markuplint/compare/@markuplint/htmx-parser@4.0.2...@markuplint/htmx-parser@4.0.3) (2024-06-25)
 
 **Note:** Version bump only for package @markuplint/htmx-parser