+++ packages/@markuplint/ml-core/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.1](https://github.com/markuplint/markuplint/compare/@markuplint/ml-core@4.10.0...@markuplint/ml-core@4.10.1) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/ml-core
+
+# [4.10.0](https://github.com/markuplint/markuplint/compare/@markuplint/ml-core@4.9.0...@markuplint/ml-core@4.10.0) (2024-10-14)
+
+### Features
+
+- **ml-core:** enabled control over parse-error output using `severity.parseError` ([7ef6d6a](https://github.com/markuplint/markuplint/commit/7ef6d6ad58845c81367d5a2944c254a12eeaa17e))
+
+# [4.9.0](https://github.com/markuplint/markuplint/compare/@markuplint/ml-core@4.8.3...@markuplint/ml-core@4.9.0) (2024-09-23)
+
+### Features
+
+- **ml-core:** update DOM API according to TypeScript DOM Libs ([b95b689](https://github.com/markuplint/markuplint/commit/b95b689a84f0a176175943edf5d4163de8b1522f))
+
+## [4.8.3](https://github.com/markuplint/markuplint/compare/@markuplint/ml-core@4.8.2...@markuplint/ml-core@4.8.3) (2024-09-02)
+
+**Note:** Version bump only for package @markuplint/ml-core
+
 ## [4.8.2](https://github.com/markuplint/markuplint/compare/@markuplint/ml-core@4.8.1...@markuplint/ml-core@4.8.2) (2024-06-25)
 
 ### Bug Fixes