+++ packages/@markuplint/shared/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.7](https://github.com/markuplint/markuplint/compare/@markuplint/shared@4.4.6...@markuplint/shared@4.4.7) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/shared
+
+## [4.4.6](https://github.com/markuplint/markuplint/compare/@markuplint/shared@4.4.5...@markuplint/shared@4.4.6) (2024-10-14)
+
+**Note:** Version bump only for package @markuplint/shared
+
+## [4.4.5](https://github.com/markuplint/markuplint/compare/@markuplint/shared@4.4.4...@markuplint/shared@4.4.5) (2024-09-23)
+
+**Note:** Version bump only for package @markuplint/shared
+
+## [4.4.4](https://github.com/markuplint/markuplint/compare/@markuplint/shared@4.4.3...@markuplint/shared@4.4.4) (2024-09-02)
+
+**Note:** Version bump only for package @markuplint/shared
+
 ## [4.4.3](https://github.com/markuplint/markuplint/compare/@markuplint/shared@4.4.2...@markuplint/shared@4.4.3) (2024-06-25)
 
 **Note:** Version bump only for package @markuplint/shared