+++ packages/@markuplint/test-tools/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.9](https://github.com/markuplint/markuplint/compare/@markuplint/test-tools@4.5.8...@markuplint/test-tools@4.5.9) (2024-10-15)
+
+**Note:** Version bump only for package @markuplint/test-tools
+
+## [4.5.8](https://github.com/markuplint/markuplint/compare/@markuplint/test-tools@4.5.7...@markuplint/test-tools@4.5.8) (2024-10-14)
+
+**Note:** Version bump only for package @markuplint/test-tools
+
+## [4.5.7](https://github.com/markuplint/markuplint/compare/@markuplint/test-tools@4.5.6...@markuplint/test-tools@4.5.7) (2024-09-23)
+
+**Note:** Version bump only for package @markuplint/test-tools
+
+## [4.5.6](https://github.com/markuplint/markuplint/compare/@markuplint/test-tools@4.5.5...@markuplint/test-tools@4.5.6) (2024-09-02)
+
+**Note:** Version bump only for package @markuplint/test-tools
+
 ## [4.5.5](https://github.com/markuplint/markuplint/compare/@markuplint/test-tools@4.5.4...@markuplint/test-tools@4.5.5) (2024-06-25)
 
 **Note:** Version bump only for package @markuplint/test-tools